Move MainActivity to presentation layer
